burme-pro/backend/requirements.txt
# Burme Editing Pro - Backend Requirements
# Built by Aung Myo Kyaw (@amkyaw.dev)

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Database (Neon PostgreSQL)
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# AI APIs
groq==0.4.2
elevenlabs==1.10.0
google-generativeai==0.3.2

# SRT/VTT Processing
srt==3.5.3

# CORS
starlette-cors==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
httpx==0.26.0

# Logging
loguru==0.7.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3